+++ packages/django-vscode/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
+## Unreleased
+
+- Report formatter failures once with a concise notification while retaining detailed diagnostics in the Django Templates output channel.
+
 See https://github.com/EpicAlbin03/django-templates-extension/releases